refactor: extract useLobbyList hook and convert lobby page to Tailwind
Move all state, data fetching, socket connection, filter sync, and
auto-refresh logic to use-lobby-list.ts. Replace ~25 inline style={{}}
objects with Tailwind utilities in page.tsx (515 lines → 145 lines of JSX).
Replace onMouseEnter/Leave hover state with Tailwind hover: utilities.
